Increase timeout to wait for interface to be up
Containers (specially those that start at the beginning of a sandbox
flavour) that wait for interfaces to be set up externally (midolman and
quagga for the bgp interfaces) fire a timeout because the provisioning
scripts are executed too late in the boot sequence. Increase this
timeout so containers do not exit too soon to reduce the likelihood of
this happening on loaded machines (e.g. CI infra).

Signed-off-by: Xavi León <xavi.leon@midokura.com>