Images: Add Openstack Neutron service to sandbox
Add the OSt Neutron service to the components in sandbox (both for kilo
and liberty). Once inside the neutron container service, both the
midonet-cli and neutron clients are available to query the virtual
topology created. Neutron requires a mysql database. For simplicity, we
reuse the official mariadb image provided by the community.

Change-Id: I92a02c26772f969eaeb0114783ca348e5028d91f
Signed-off-by: Xavi León <xavi.leon@midokura.com>