Failfast on sandbox commands
Sandbox currently does not fail when building a base image and thus it
tries to build the child image regardless. Make evident that a command
failed by also return a positive integer a la POSIX to show an error so
it can be included in other scripts relying on proper exit status
handling (0 for normal exit status, 1 for error). We also skip images
in flavour definitions that are not meant to be built (just pulled) like
mariadb or elk containers.

Change-Id: I7db0296ffd6d02842fec9b14be91f33a92ef44ac
Signed-off-by: Xavi León <xavi.leon@midokura.com>