Fake Snort, compiled implementation
The original fake snort, built in python did not set the read
timeout for pcap, so on some systems would delay for a long time before
bouncing packets. This prevented the tests from passing.

The pcap_set_timeout call is not available in the python pcap wrapper,
so it was not possible to just add a call to it. This meant I needed to
reimplement the whole thing in C, just so I could make that call. This
solves the failing test issue though.

Change-Id: Id496a59cf9b029bbf5c5706256b0c99d2b017779