Images: deterministically set ZK_IDs
ZK IDs were not deterministically assigned during ZK start up script, so
they conflicted with the one specified in the sandbox flavour file. The
fix consists in assign ZK_IDs based on the last octet of the ip
address allocated by docker (no collision as we are using /24 for the
sandbox network).

Change-Id: I1023f3ecffc37ca990460d254ddcb6b85bd03c79
Signed-off-by: Xavi León <xavi.leon@midokura.com>