Update midonet* images to use the unstable packages for master
Now that we have a nightly pipeline that publishes unstables packages,
we can create midonet* (cluster and midolman) containers with those
packages. Of course, packages can be reinstalled on a per-case basis
using the overrides features as we currently do in the midonet gates.
This change will allow spawning a complete multinode environment with
the latest unstables packages. Also changed from zulu8 to openjdk ppa
repo as the former failed quite often due to issues with their CDN.

Change-Id: I3a0421cbc262c326bd364cc778518d451f676cc2
Signed-off-by: Xavi León <xavi.leon@midokura.com>