VPNaaS: add configuration on base to allow overrides
We support neutron advanced services on different OSt distributions
depending on customers. However, we'd like the default configuration
to work with our upstream vanilla plugin. This patch reorganizes
the neutron structure a bit to provide advanced services
configuration on the base image to allow activating them on the
configuration file on overrides for specific deployments.

Change-Id: I3be4645f656f92f5de38ff6968b52f1a90ee2075
Signed-off-by: Xavi León <xavi.leon@midokura.com>