Add env var ZOO_IN_MEM and install openjdk-7-jdk
In the ZK flavour environment definition, ZOO_IN_MEM allows to select whether
to keep ZK in disk or in memory. By default it will keep it in memory if it is
unset or true. Only if it is set to false it will be kept in disk.
Openjdk-7 in ZK container is needed for Jolokia JVM monitoring purposes.

Change-Id: I043faf08055f71794a3594ed06e42af15f6364c0
Signed-off-by: Lucas Eznarriaga <lucas.ez@gmail.com>