Neutron: create docker image for FJ plugin version
For running our VPNaaS tests and other neutron based tests, we need
a specific version of our neutron plugin. To do this, we had an
override on different repos that downloaded that specific version.
This has proven to be flaky as it sometimes downloading the package
from artifactory takes a lot of time, delaying the creation of the
container, and thus failing the checks we perform before starting
the tests.

This patch creates a FJ specific image to be pushed to artifactory
so we don't need to have specific overrides for neutron, removing
the timeout issue when spinning up the container.

Ref.: MI-700
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: I57e69c004d0adf4c5b77aa1448467186d310e5fd