Fixing recursive-include calls in manifest
When executing pip install it complained about
these calls stating it should follow the below syntax:
recursive-include <folder> <pattern1> <pattern2> ...

The issue was that the assets were not installed and the images
specified in the 1.9 flavours could not be built because the
base images couldn't be found.

With this fix the base images and the composer folder get installed
properly and you can then build the 1.9 flavours without problems.

Change-Id: I4b7c780253288ec376851dc8ddcdfaeafc94cf9d
Signed-off-by: Nazim Saouli <nazim@midokura.com>