Add the Kubernetes and Raven services
To test our integration with container orchestrators, prepare an image
to run the K8s services and another one with the Raven service. The way
they deploy is slightly different from a regular sandbox, since the K8s
services run at host network namespace and midolman as well.

A future improvement, running MidolMan in a container inside the Kubelet
one and do kind of Docker-in-Docker there it is proposed, but right now
we are happy with this single-host approach.

To understand better what are the differences between a regular sandbox
development and this one, check out the document:

https://docs.google.com/document/d/1nrFkWJae9PkfLPSyX3oSupN3XuersVcdFAmgR1YJPrA/edit
Signed-off-by: Jaume Devesa <devvesa@gmail.com>

Change-Id: I4b409a6d48d1c78a25d33485d6904cafd93eca65