Quagga: add more logs and allow passive speakers
We try to connect between bgp speakers every few seconds (even if the
bgp neighbor is not set up on the other end). When both ends are up and
running, it may be the case that both try to establish a connection
at the same time creating a connection collision and closing the
connection of the bgp speaker with a higher router id (in our case, the
midolman bgp speaker). Once this happens, apparently the external bgp
speaker does not try to reconnect and the connection is never
established.

This patch allows a bgp speaker to be specified as passive, so
connections are only initiated by midolman when we set up the virtual
bgp router.

This patch also adds more logging on the external bgp speakers and mounts
the log directory in memory to avoid hitting disk every second when
recording the result of the heartbeats with all peering bgps.

Change-Id: I15e18c831a248ee9e0520f1fe197ee8a25b06c9d
Signed-off-by: Xavi León <xavi.leon@midokura.com>