Kuryr: install with pip
PBR depends on some pip magic and for this reason we should be using
its latest version to install kuryr.

We get directly the version from pypi using get_pip.py to avoid the
distro setuptools and pip from conflicting with the latest we have to
upgrade to.

Using pip3 installs the requirements for you, so we save one build
step!

Change-Id: Ifea5063e4f0cc714c08a4c525044e9f1ee4eee54
Signed-off-by: Antoni S. Puimedon <antonisp@celebdor.com>