Add --no-install-recommends installing midolman
A recent change on one of the grub packages only installed as a
recommended package causes an error installing midolman because it asks
for a configuration option in a non-interactive console.

Since the package it is not needed at all, let's skip its installation
by using --no-install-recommends.

It also saves space, which is good.

Change-Id: Id1f794c4b23334b7469c8c7b1a8791edb8311f0c