Allow provision and override specs in flavour file
This patch allows the specification of the override folder and provision
scripts in the flavour yaml. To specify this, the user only needs to add
the following to the flavour yaml:

provision: relative/path/to/provisioning_script.sh

override: relative/override/folder

The provision and override paths will be overridden by the values given
with the command line if any.

Ref.: MI-1554 MI-1555
Change-Id: Id1e0d9a9090b9db6d17c875724cd8b82344f3503
Signed-off-by: Xavi León <xavi.leon@midokura.com>