Add force option to build command
Until now, the build command forced building an image. This patch adds
an option --force to force building the image if it exists (not forced
by default now as in the case of build_all command).

This patch also passes the force flag from the build_all command to the
build_command so the base images are re-built also depending on that
flag.

Ref: IIOT-448
Change-Id: I1212c7d9d6d8bb5120850b633745796be7e449a3
Signed-off-by: Xavi León <xavi.leon@midokura.com>